Add initial files
This is a little project experimenting with an idea for viewing and
interacting with many article-like pages of content like you would
records in a record shop, a la enid.fm.

This commit adds the svelte project as it currently stands. I have a
basic demo setup, with a functioning 'browse' feature, letting you hover
over the records and see their preview content.