Add 'focus' state for records when clicked on
Clicking on records now presents them in a 'focus' view, displaying
its full content, with all of the other records pushed off-screen.

A bit of work still needs to be done with the 'un-focusing' behaviour.
Currently, unfocusing will put you back at container.scrollTop == 0;
which is a bit jarring after focusing and unfocusing a record from
far down the list.